+++ miden-crypto/src/hash/mod.rs
@@ -2,9 +2,13 @@
 
 use super::{CubeExtension, Felt, FieldElement, StarkField, Word, ZERO};
 
-/// Blake2s hash function.
+/// Blake3 hash function.
 pub mod blake;
 
+/// Keccak hash function.
+pub mod keccak;
+
+/// Poseidon2 hash function.
 pub mod poseidon2 {
     pub use super::algebraic_sponge::poseidon2::Poseidon2;
 }