+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [BREAKING] Rename `SparseValuePath` to `SimpleSmtProof` ([#477](https://github.com/0xMiden/crypto/pull/477)).
 - Validate `NodeIndex` depth ([#482](https://github.com/0xMiden/crypto/pull/482)).
 - [BREAKING] Rename `ValuePath` to `MerkleProof` ([#483](https://github.com/0xMiden/crypto/pull/483)).
+- Added an implementation of Keccak256 hash function ([#487](https://github.com/0xMiden/crypto/pull/487)).
 
 # 0.15.9 (2025-07-24)